--- a/dataprofiler/tests/space_time_analysis/dataset_generation.py
+++ b/dataprofiler/tests/space_time_analysis/dataset_generation.py
@@ -16,7 +16,7 @@ except ImportError:
     import dataprofiler as dp
 
 
-class NpEncoder(json.JSONEncoder):
+class NumpyEncoder(json.JSONEncoder):
     def default(self, obj):
         if isinstance(obj, np.integer):
             return int(obj)